Fixed bugs in displaying fight logs; Refactored database

+++ FarmHeroes/Data/FarmHeroes.Data.Models/HeroModels/Health.cs
@@ -8,9 +8,9 @@
 
         public int Id { get; set; }
 
-        public long Current { get; set; } = InitialHealth;
+        public int Current { get; set; } = InitialHealth;
 
-        public long Maximum { get; set; } = InitialHealth;
+        public int Maximum { get; set; } = InitialHealth;
 
         public virtual Hero Hero { get; set; }
     }
